Make it easier to build
I cleaned up a bit compilation tasks.
- Merged both 'inject-player' and 'inject-index' into 'inject-html'.
- Added a 'compile' task

Added tasks to help install & build
- build handles ffmpeg lib
- run launches the app

Also updates README to reflect this changes.

Also added nodewebkit as a dependency in package.json.
It is a dependency after all.

Please test the build and run scripts on different platforms. I tested
it on Windows and it works just fine.